traditional_rag/requirements.txt
ollama
faiss-cpu
numpy
sentence-transformers
